+++ src/App/Markdown.fs
@@ -12,7 +12,20 @@ let style = [
     rule "h5" [ Css.allRevert ]
     rule "ul" [ Css.allRevert ]
     rule "li" [ Css.allRevert ]
-    rule "p" [ Css.marginTop (em 1); Css.marginBottom (em 1)]
+    rule "p" [
+        Css.marginTop (em 1);
+        Css.marginBottom (em 1)
+    ]
+    rule "table p" [
+        Css.marginTop (em 0);
+        Css.marginBottom (em 0)
+    ]
+    rule "code" [
+        Css.fontSize (percent 80)
+    ]
+    rule "pre" [
+        Css.padding (rem 0.5)
+    ]
     rule "code.more" [
         Css.maxHeight (rem 12)
         Css.overflowYHidden
@@ -21,9 +34,9 @@ let style = [
         Css.transition "max-height 0.5s cubic-bezier(0, 1, 0, 1)" // the transition to this state
      ]
     rule "code.full" [
-        Css.maxHeight (px 1000)
-        Css.transition "max-height 0.5s cubic-bezier(0, 1, 0, 1)" // the transition to this state
-        //Css.transition "max-height 1s ease-in-out"
+        Css.maxHeight (px 1500)
+        //Css.transition "max-height 0.5s cubic-bezier(0, 1, 0, 1)" // the transition to this state
+        Css.transition "max-height 1s ease-in-out"
         //Css.transition "-webkit-mask-image 4s ease-in-out"
         //Css.transition "mask-image 4s ease-in-out"
         Css.custom("-webkit-mask-image", "linear-gradient(to bottom, black 100%, black 100%)")