+++ src/Sutil/Interop.fs
@@ -22,7 +22,7 @@ let set<'T> (ob:obj) (name:string) (value:'T) : unit = jsNative
 let get<'T> (ob:obj) (name:string) : 'T = jsNative
 
 [<Emit("delete $0[$1]")>]
-let delete ob name = jsNative
+let delete ob name : unit = jsNative
 
 [<Emit("$0.hasOwnProperty($1)")>]
 let exists (ob:obj) (name:string) : bool= jsNative