+++ tests/Test.Store.fs
@@ -10,7 +10,7 @@ let tests = testList "Sutil.Store" [
         let mutable v = 0
         let u = s |> Store.subscribe (fun v' -> v <- v')
         u.Dispose()
-        Expect.areEqual v 42
+        Expect.areEqual(v,42)
 
     testCase "Store updates when no subscribers" <| fun () ->
         let s = Store.make 42
@@ -19,18 +19,18 @@ let tests = testList "Sutil.Store" [
         let mutable v = 0
         let u = s |> Store.subscribe (fun v' -> v <- v')
         u.Dispose()
-        Expect.areEqual v 43
+        Expect.areEqual(v,43)
 
     testCase "Update store notifies subscriber" <| fun () ->
         let s = Store.make 42
         let mutable v = 0
         let u = s |> Store.subscribe (fun v' -> v <- v')
-        Expect.areEqual v 42
+        Expect.areEqual(v,42)
 
         Store.set s 43
 
         u.Dispose()
-        Expect.areEqual v 43
+        Expect.areEqual(v,43)
 
     testCase "Multiple subscribers initialize" <| fun () ->
         let s = Store.make 42
@@ -41,8 +41,8 @@ let tests = testList "Sutil.Store" [
         let u1 = s |> Store.subscribe (fun v -> v1 <- v)
         let u2 = s |> Store.subscribe (fun v -> v2 <- v)
 
-        Expect.areEqual v1 42
-        Expect.areEqual v2 42
+        Expect.areEqual(v1,42)
+        Expect.areEqual(v2,42)
 
     testCase "Multiple subscribers update" <| fun () ->
         let s = Store.make 42
@@ -52,14 +52,13 @@ let tests = testList "Sutil.Store" [
 
         let u1 = s |> Store.subscribe (fun v -> v1 <- v)
         let u2 = s |> Store.subscribe (fun v -> v2 <- v)
-
-        Expect.areEqual v1 42
-        Expect.areEqual v2 42
+        Expect.areEqual(v1,42)
+        Expect.areEqual(v2,42)
 
         Store.set s 43
 
-        Expect.areEqual v1 43
-        Expect.areEqual v2 43
+        Expect.areEqual(v1,43)
+        Expect.areEqual(v2,43)
 
     testCase "Dispose terminates subscription" <| fun () ->
         let s = Store.make 42
@@ -70,15 +69,15 @@ let tests = testList "Sutil.Store" [
         let u1 = s |> Store.subscribe (fun v -> v1 <- v)
         let u2 = s |> Store.subscribe (fun v -> v2 <- v)
 
-        Expect.areEqual v1 42
-        Expect.areEqual v2 42
+        Expect.areEqual(v1,42)
+        Expect.areEqual(v2,42)
 
         u1.Dispose();
 
         Store.set s 43
 
-        Expect.areEqual v1 42
-        Expect.areEqual v2 43
+        Expect.areEqual(v1,42)
+        Expect.areEqual(v2,43)
 
 
     testCase "No dispose on number of subscribers becoming 0" <| fun () ->
@@ -96,8 +95,8 @@ let tests = testList "Sutil.Store" [
         u |> ignore
         Store.set s 43
 
-        Expect.areEqual v1 42
-        Expect.areEqual v2 43
+        Expect.areEqual(v1,42)
+        Expect.areEqual(v2,43)
 
     testCase "Notify all updates even when equal" <| fun () ->
         let inputValue = 42
@@ -113,6 +112,6 @@ let tests = testList "Sutil.Store" [
 
         u.Dispose()
 
-        Expect.areEqual n 3
+        Expect.areEqual(n,3)
 
 ]