+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 1.0.0-beta-007
+
+- Support for adopted stylesheets
+- Fixes for disposal of bind nodes
 ### 1.0.0-beta-006
 
 - Eliminate use of eval in webcomponent.js